calibrage

+++ PoteriePerso/VariablesGlobales.h
@@ -8,4 +8,4 @@
 extern CPoterieSequence *seq;
 
 //Echelle permettant la correspondance echelle/cm
-extern float echelle;
+extern double echelle;
